requirements.txt
Django==4.2.17
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.2
pandas==2.2.3
stripe==11.5.0